update direction slice & containers
